+++ comics/comics/pondusbt.py
@@ -5,14 +5,14 @@ from comics.comics.pondus import Meta as PondusMeta
 
 class Meta(PondusMeta):
     name = 'Pondus (bt.no)'
-    url = 'http://www.bt.no/tegneserier/?type=pondus'
+    url = 'http://www.bt.no/bergenpuls/tegneserier/tegneserier_pondus/'
 
 class Crawler(CrawlerBase):
-    history_capable_days = 14
+    history_capable_days = 32
     schedule = 'Mo,Tu,We,Th,Fr,Sa'
     time_zone = 1
 
     def crawl(self, pub_date):
-        url = 'http://images.bt.no/gfx/cartoons/pondus/%s.gif' % (
+        url = 'http://www.bt.no/external/cartoon/pondus/%s.gif' % (
             pub_date.strftime('%d%m%y'),)
         return CrawlerImage(url)