+++ comics/comics/rockybt.py
@@ -3,7 +3,7 @@ from comics.comics.rocky import Meta as RockyMeta
 
 class Meta(RockyMeta):
     name = 'Rocky (bt.no)'
-    url = 'http://www.bt.no/tegneserier/rocky/'
+    url = 'http://www.bt.no/bergenpuls/tegneserier/tegneserier_rocky/'
 
 class Crawler(CrawlerBase):
     history_capable_days = 162
@@ -11,6 +11,6 @@ class Crawler(CrawlerBase):
     time_zone = 1
 
     def crawl(self, pub_date):
-        url = 'http://images.bt.no/gfx/cartoons/rocky/%s.gif' % (
+        url = 'http://www.bt.no/external/cartoon/rocky/%s.gif' % (
             pub_date.strftime('%d%m%y'),)
         return CrawlerImage(url)