+++ docs/changes.rst
@@ -11,6 +11,7 @@ In development
 
 **Crawlers**
 
+- New: ``fagprat``
 - New: ``satw``
 - Update ``m`` as it moved from start.no to dagbladet.no.
 - Update ``treadingground`` to not crash if URL is not found.