+++ docs/changes.rst
@@ -13,6 +13,7 @@ In development
 
 - New: ``fagprat``
 - New: ``satw``
+- New: ``stickycomics``
 - Update ``bizarro`` which moved to a new site.
 - Update ``lookingforgroup`` after feed change.
 - Update ``m`` as it moved from start.no to dagbladet.no.