+++ docs/changes.rst
@@ -18,6 +18,7 @@ In development
 - Update ``bizarro`` which moved to a new site.
 - Update ``lookingforgroup`` after feed change.
 - Update ``m`` as it moved from start.no to dagbladet.no.
+- Update ``nerfnow`` which crashed when it did not find an image URL.
 - Update ``treadingground`` to not crash if URL is not found.