+++ docs/changes.rst
@@ -17,6 +17,7 @@ In development
 - New: ``tehgladiators``
 - Update ``bizarro`` which moved to a new site.
 - Update ``countyoursheep`` after URL changes.
+- Update ``darylcagle`` after change from GIF to JPEG.
 - Update ``lookingforgroup`` after feed change.
 - Update ``m`` as it moved from start.no to dagbladet.no.
 - Update ``nerfnow`` which crashed when it did not find an image URL.